+++ app/src/main/java/com/wamcstudios/moviesapp/home/data/local/entity/MediaEntity.kt
@@ -53,7 +53,7 @@ data class MediaEntity(
     val imdbIdDetail: String,
     val productionCompaniesDetail: List<ProductionCompany>,
     val productionCountriesDetail: List<ProductionCountry>,
-    val revenueDetail: Int,
+    val revenueDetail: Long,
     val runtimeDetail: Int,
     val spokenLanguageDetail: List<SpokenLanguage>,
     val statusDetail: String,
