+++ app/src/main/java/com/wamcstudios/moviesapp/home/data/remote/dto/MediaDto.kt
@@ -74,7 +74,7 @@ data class MediaDto(
     @SerializedName("production_countries")
     val productionCountriesDetail: List<ProductionCountryDto>?,
     @SerializedName("revenue")
-    val revenueDetail: Int?,
+    val revenueDetail: Long?,
     @SerializedName("runtime")
     val runtimeDetail: Int?,
     @SerializedName("spoken_languages")
